use basic entity
