+++ custom_components/husqvarna_automower/config_flow.py
@@ -1,9 +1,9 @@
 """Config flow to add the integration via the UI."""
 import logging
 
-import voluptuous as vol
 from aioautomower import GetAccessToken, GetMowerData, TokenError
 from aiohttp.client_exceptions import ClientConnectorError, ClientResponseError
+import voluptuous as vol
 
 from homeassistant import config_entries
 from homeassistant.const import (
